Cast mr_bulk_load_edges timeout to bigint, add m0012, bump to 0.6.8
- Use 120000::bigint so function signature matches extension (timeout_msec bigint)
- Add migration m0012 for already-deployed DBs
- Bump tentura_server to 0.6.8

Made-with: Cursor

+++ packages/server/lib/data/database/migration/m0011.dart
@@ -83,7 +83,7 @@ BEGIN
 
   _total := _total + _edge_count;
 
-  PERFORM mr_bulk_load_edges(_src, _dst, _weight, _magnitude, _context, 120000);
+  PERFORM mr_bulk_load_edges(_src, _dst, _weight, _magnitude, _context, 120000::bigint);
 
   -- Read Updates Filters
   SELECT _total + count(*)::int INTO _total
